pyproject.toml
# Poetry

[tool.poetry]
name = "the-spymaster-bot"
package-mode = false
description = "Telegram bot service for Codenames board game."
authors = ["Asaf Kali <asaf.kali@mail.huji.ac.il>"]
readme = "README.md"

[tool.poetry.dependencies]
# Core
python = "^3.8"
codenames = "^4.0.3"
the-spymaster-util = { version = "~3.2", extras = ["all"] }
the-spymaster-api = "^2.1"
# Telegram bot
python-telegram-bot = "^13.11"
# Models
pydantic = "^1.9"
pynamodb = "^5.4.1"
# Utils
requests = "^2.26.0"
cachetools = "^4.2.2"
urllib3 = "<2.0" # Specified due to dependency resolution issue
# Monitor
sentry-sdk = "^1.19.1"

[tool.poetry.group.security.dependencies]
tornado = "^6.3.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.1"
parameterized = "^0.8.1"
freezegun = "^1.2.1"
coverage = "^6.3.2"

[tool.poetry.group.lint.dependencies]
ruff = "^0.7.1"
black = "^24.3"
isort = "^5.11.4"
mypy = "^0.991"
types-requests = "^2.28.11.8"
types-cachetools = "^5.2.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Test

[tool.pytest.ini_options]
pythonpath = ["app"]
addopts = "-s --disable-pytest-warnings"

[tool.coverage.run]
source = ["app"]

# Lint

[tool.ruff]
line-length = 120
exclude = ["local", "tf", ".deployment"]

[tool.ruff.lint]
select = ["B", "C", "E", "F", "W"]
ignore = []

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
skip = ["__init__.py", "tf/", ".deployment/"]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true
exclude = ["local/", "tf/", "dist/", "build/", ".deployment/"]
